Fixes based on first review
